More general DOI updating setup

+++ run_caltechdata_checks.py
@@ -8,7 +8,8 @@ if os.path.isdir('data') == False:
     os.mkdir('data')
 os.chdir('data')
 
-shutil.rmtree(collection)
+if os.path.isdir(collection) == True:
+    shutil.rmtree(collection)
 
 token = os.environ['TINDTOK']
 
@@ -16,5 +17,8 @@ get_caltechdata(collection,datacite=True)
 fix_multiple_links(collection,token)
 check = input("Do you want to update DOI metadata with DataCite? Y or N:")
 if check == 'Y':
-    password = getpass.getpass("Enter the TIND DataCite Password:")
-    update_datacite_metadata(collection,token,password)
+    username = getpass.getpass("Enter the DataCite Username:")
+    password = getpass.getpass("Enter the DataCite Password:")
+    prefix = getpass.getpass("Enter the DataCite Prefix:")
+    access = [{'username':username,'password':password,'prefix':prefix}]
+    update_datacite_metadata(collection,token,access)
